add finale modification

+++ route/codepromoRoute.js
@@ -33,27 +33,59 @@ router.post('/addcodepromo', (req, res) => {
     });
   });
 
-  router.delete('/deletecodepromo/:id', (req, res) => {
-    const promoId = req.params.id;
-    
+
+  router.delete('/deletecodepromo', (req, res) => {
+    const promoId = req.body.id;
+  
+    if (!promoId) {
+      res.status(400).json({ error: 'Missing promo ID in request body' });
+      return;
+    }
+  
     const sql = 'DELETE FROM codepromo WHERE id = ?';
     const values = [promoId];
-    
+  
     codepromo.query(sql, values, (err, result) => {
       if (err) {
         console.error('Erreur lors de la suppression du code promo :', err);
         res.status(500).json({ error: 'Erreur lors de la suppression du code promo' });
         return;
       }
-      
+  
       if (result.affectedRows === 0) {
         res.status(404).json({ error: 'Code promo non trouvé' });
         return;
       }
-      
+  
       res.json({ message: 'Code promo supprimé avec succès' });
     });
   });
+
+
+  const cors = require('cors');
+  
+  router.post('/codpromo', (req, res) => {
+    const code = req.body.code;
+  
+    // Query the database to check if the promotion code exists
+    codepromo.query('SELECT * FROM codepromo WHERE code = ?', [code], (err, results) => {
+      if (err) {
+        console.error('Error querying MySQL:', err);
+        res.status(500).json({ error: 'An internal server error occurred' });
+        return;
+      }
+  
+      if (results.length === 0) {
+        // Promotion code doesn't exist
+        res.json('Promotion code not found' );
+      } else {
+        // Promotion code exists, retrieve the discount percentage
+        const discountPercentage = results[0].discount;
+        res.json({ discountPercentage });
+      }
+    });
+  });
+
